requirements.txt
certifi==2017.7.27.1
cffi==1.10.0
chaussette==1.3.0
chardet==3.0.4
circus==0.14.0
circus-web==1.0.0
coreapi==2.3.1
coreschema==0.0.4
cryptography==1.7.2
Django==1.11.4
django-cors-headers==2.1.0
django-guardian==1.4.9
django-import-export=1.1.0
django-nested-admin==3.0.20
django-object-actions==1.0.0
django-report-builder==6.0.2
django-rest-framework==0.1.0
django-rest-swagger==2.1.2
django-reversion==3.0.1
django-reversion-compare==0.8.5
django-templated-mail==1.1.1
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
djangorestframework-recursive==0.1.2
djoser==1.1.5
drf-nested-routers==0.90.0
gevent==1.2.2
idna==2.6
itypes==1.1.0
Jinja2==2.9.6
Markdown==2.6.8
MarkupSafe==1.0
openapi-codec==1.3.2
postmarker==0.11.3
pyasn1==0.3.2
pycparser==2.18
PyJWT==1.5.3
python-monkey-business==1.0.0
pytz==2017.2
PyYAML==3.12
requests==2.18.4
simplejson==3.11.1
six==1.10.0
tornado==4.5.3
uritemplate==3.0.0
urllib3==1.22
